move estimates from measures -> algorithms
